backend/requirements.txt
pandas>=2.2.0
openpyxl>=3.1.5
google-auth>=2.30.0
google-auth-oauthlib>=1.2.0
requests>=2.32.0
pyyaml>=6.0
